@@ -526,7 +526,7 @@
     "@types/node": "^6.0.40",
     "mocha": "^3.2.0",
     "tslint": "^4.2.0",
-    "typescript": "^2.6.2",
+    "typescript": "^2.8.3",
     "vscode": "^1.0.0"
   },
   "dependencies": {
